Pokeball almost done

+++ Base/Source/Enemy_Poison.h
@@ -8,7 +8,7 @@ class Enemy_Poison
 public:
 	static Enemy_Poison* pointer(){ return m_pointer; };
 	void Init();
-	void Update(float dt, Map* map);
+	void Update(double dt, Map* map);
 	void render();
 	void Exit();
 	Vector3 Poison(Vector3 Movement);
